Remove ArcCell, MsQueue, TreiberStack
