+++ src/lib.rs
@@ -1,14 +1,13 @@
 //! Tools for concurrent programming.
 //!
 //! * Atomics
-//!     * [`ArcCell<T>`] is a shared mutable [`Arc<T>`] pointer.
 //!     * [`AtomicCell<T>`] is equivalent to [`Cell<T>`], except it is also thread-safe.
 //!     * [`AtomicConsume`] allows reading from primitive atomic types with "consume" ordering.
 //!
 //! * Data structures
 //!     * [`deque`] module contains work-stealing deques for building task schedulers.
-//!     * [`MsQueue<T>`] and [`SegQueue<T>`] are simple concurrent queues.
-//!     * [`TreiberStack<T>`] is a lock-free stack.
+//!     * [`ArrayQueue<T>`] is a bounded MPMC queue.
+//!     * [`SegQueue<T>`] is an unbounded MPMC queue.
 //!
 //! * Thread synchronization
 //!     * [`channel`] module contains multi-producer multi-consumer channels for message passing.
@@ -22,15 +21,13 @@
 //!     * [`CachePadded<T>`] pads and aligns a value to the length of a cache line.
 //!     * [`scope()`] can spawn threads that borrow local variables from the stack.
 //!
-//! [`ArcCell<T>`]: atomic/struct.ArcCell.html
 //! [`Arc<T>`]: https://doc.rust-lang.org/std/sync/struct.Arc.html
 //! [`AtomicCell<T>`]: atomic/struct.AtomicCell.html
 //! [`Cell<T>`]: https://doc.rust-lang.org/std/cell/struct.Cell.html
 //! [`AtomicConsume`]: atomic/trait.AtomicConsume.html
 //! [`deque`]: deque/index.html
-//! [`MsQueue<T>`]: queue/struct.MsQueue.html
+//! [`ArrayQueue<T>`]: queue/struct.ArrayQueue.html
 //! [`SegQueue<T>`]: queue/struct.SegQueue.html
-//! [`TreiberStack<T>`]: stack/struct.TreiberStack.html
 //! [`channel`]: channel/index.html
 //! [`ShardedLock<T>`]: sync/struct.ShardedLock.html
 //! [`RwLock<T>`]: https://doc.rust-lang.org/std/sync/struct.RwLock.html
@@ -68,13 +65,10 @@ mod _epoch {
 #[doc(inline)]
 pub use _epoch::crossbeam_epoch as epoch;
 
-mod arc_cell;
-
 extern crate crossbeam_utils;
 
 /// Atomic types.
 pub mod atomic {
-    pub use arc_cell::ArcCell;
     pub use crossbeam_utils::atomic::AtomicCell;
     pub use crossbeam_utils::atomic::AtomicConsume;
 }
@@ -114,16 +108,10 @@ cfg_if! {
         extern crate parking_lot;
 
         mod sharded_lock;
-        mod treiber_stack;
         mod wait_group;
 
         /// Concurrent queues.
         pub mod queue;
-
-        /// Concurrent stacks.
-        pub mod stack {
-            pub use treiber_stack::TreiberStack;
-        }
 
         /// Thread synchronization primitives.
         pub mod sync {
